requirements.txt
Django==4.1.5
djangorestframework==3.14.0
celery==5.2.7
requests==2.28.2
python-dotenv==0.21.0
pypdf==3.4.0
gunicorn==20.1.0
psycopg2==2.9.5
redis==4.5.1